mqtt clientid done right (+no space in id)

+++ RX_FSK/src/mqtt.cpp
@@ -27,7 +27,7 @@ void MQTT::init(const char* host, uint16_t port, const char* id, const char *use
     
     Serial.println("[MQTT] pubsub client");
     mqttClient.setServer(ip, port);
-    snprintf(buffer, 20, "%s%6ld", id, random(0, 1000));
+    snprintf(buffer, 20, "%s%04d", id, (int)random(0, 1000));
     buffer[20] = 0;
     mqttClient.setClientId(buffer);
     if (strlen(password) > 0) {
